requirements.txt
psutil
pyyaml
python-dotenv
redis
rel
dotenv
python-snap7==1.3